move contacts in marketing_campaigns
